@@ -4,11 +4,12 @@
   "description": "Adds an admin section to the hexo server",
   "main": "index.js",
   "scripts": {
-    "prepublish": "gulp build"
+    "start": "webpack-dev-server --mode development --open --hot",
+    "build": "./node_modules/.bin/webpack --mode production"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/jaredly/hexo-admin.git"
+    "url": "https://github.com/tianyingchun/hexo-admin-tm.git"
   },
   "keywords": [
     "hexo",
@@ -25,40 +26,56 @@
   "dependencies": {
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.5.0",
-    "browserify": "^11.2.0",
     "connect-auth": "^0.6.1",
     "cookie-parser": "^1.3.4",
-    "deep-assign": "^2.0.0",
+    "deep-assign": "^3.0.0",
     "express-session": "^1.10.3",
     "extend": "^3.0.0",
     "hexo-front-matter": "^0.2.2",
-    "hexo-fs": "0.2.3",
-    "hexo-util": "^0.1.6",
-    "less": "^2.5.1",
+    "hexo-fs": "1.0.1",
+    "hexo-util": "^0.6.3",
     "moment": "^2.7.0",
     "serve-static": "^1.4.0",
-    "write-good": "^0.10.0"
+    "write-good": "^0.13.1"
   },
   "devDependencies": {
-    "browserify": "^11.2.0",
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-decorators": "^7.0.0",
+    "@babel/polyfill": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.0.0",
+    "@types/node": "^10.12.10",
+    "@types/react": "^16.4.7",
+    "@types/react-dom": "^16.0.6",
+    "@types/react-router-dom": "^4.3.0",
+    "autoprefixer": "^9.0.2",
+    "babel-loader": "^8.0.4",
+    "classnames": "^2.2.6",
+    "clean-webpack-plugin": "^1.0.0",
     "codemirror": "^5.24.2",
-    "es6-promise": "^1.0.0",
-    "gulp": "^3.9.1",
-    "gulp-less": "^3.0.5",
-    "gulp-rename": "^1.2.2",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-util": "^3.0.7",
-    "less": "^2.6.1",
-    "lodash": "^2.4.1",
-    "marked": "^0.3.2",
-    "moment": "^2.7.0",
-    "react": "^0.11.1",
-    "react-router": "^0.4.1",
-    "reactify": "^0.14.0",
-    "superagent": "^0.18.2",
-    "superagent-browserify": "0.0.2",
-    "vinyl-buffer": "^1.0.0",
-    "vinyl-source-stream": "^1.1.0",
-    "watchify": "^3.7.0"
+    "css-loader": "^1.0.1",
+    "eslint": "^5.9.0",
+    "file-loader": "^2.0.0",
+    "fork-ts-checker-webpack-plugin": "^0.5.0",
+    "html-webpack-plugin": "^3.2.0",
+    "less": "^3.9.0",
+    "less-loader": "^4.1.0",
+    "lodash": "^4.17.11",
+    "marked": "^0.5.2",
+    "postcss-loader": "^3.0.0",
+    "prop-types": "^15.6.2",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-hot-loader": "^4.3.12",
+    "react-router-dom": "^4.3.1",
+    "style-loader": "^0.23.1",
+    "superagent": "^4.0.0",
+    "tslint": "^5.11.0",
+    "typescript": "^3.2.1",
+    "webpack": "^4.14.0",
+    "webpack-cli": "^3.0.8",
+    "webpack-dev-server": "^3.1.4"
   }
 }
