@@ -1,12 +1,12 @@
 
-var Link = require('react-router').Link
-var React = require('react')
+const React = require('react')
+const { Link } = require('react-router-dom')
 
-var App = React.createClass({
-  render: function () {
+class App extends React.Component {
+  render() {
     return <div className="app">
       <div className="app_header">
-        <img src="logo.png" className="app_logo"/>
+        <img src="logo.png" className="app_logo" />
         <span className="app_title">Hexo Admin</span>
         <ul className="app_nav">
           <li><Link to="posts">Posts</Link></li>
@@ -17,10 +17,10 @@ var App = React.createClass({
         </ul>
       </div>
       <div className="app_main">
-        <this.props.activeRouteHandler/>
+        {this.props.children}
       </div>
     </div>;
   }
-})
+}
 
 module.exports = App
