@@ -1,54 +1,46 @@
 
-var React = require('react/addons')
-var PT = React.PropTypes
+var React = require('react')
+var PT = require('prop-types')
 var api = require('./api')
 
-var SettingsCheckbox = React.createClass({
-  propTypes: {
-    name: PT.string.isRequired,
-    label: PT.string.isRequired,
-    style: PT.object,
-    enableOptions: PT.object,
-    disableOptions: PT.object,
-    onClick: PT.func
-  },
-
-  getInitialState: function () {
-    return {
+class SettingsCheckbox extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
       checked: false,
     }
-  },
+  }
 
-  componentDidMount: function() {
+  componentDidMount() {
     var name = this.props.name
-    api.settings().then( (settings) => {
+    api.settings().then((settings) => {
       var checked
       if (!settings.options) {
         checked = false
       } else {
         checked = !!settings.options[name]
       }
-      this.setState({checked: checked})
+      this.setState({ checked: checked })
     })
-  },
+  }
 
-  handleChange: function(e) {
+  handleChange(e) {
     var name = this.props.name
     var addedOptions = e.target.checked ? this.props.enableOptions
-                                        : this.props.disableOptions
+      : this.props.disableOptions
     var value = e.target.checked
-    api.setSetting(name, value, addedOptions).then( (result) => {
+    api.setSetting(name, value, addedOptions).then((result) => {
       console.log(result.updated)
       this.setState({
         checked: result.settings.options[name]
       });
     });
-  },
+  }
 
-  render: function() {
+  render() {
     return (
       <p style={this.props.style}>
-      <label>
+        <label>
           <input
             checked={this.state.checked}
             type="checkbox"
@@ -60,6 +52,16 @@ var SettingsCheckbox = React.createClass({
       </p>
     );
   }
-});
+}
+
+SettingsCheckbox.propTypes = {
+  name: PT.string.isRequired,
+  label: PT.string.isRequired,
+  style: PT.object,
+  enableOptions: PT.object,
+  disableOptions: PT.object,
+  onClick: PT.func
+};
+
 
 module.exports = SettingsCheckbox
