@@ -1,62 +1,62 @@
 
-var React = require('react/addons')
-var PT = React.PropTypes
+var React = require('react')
+var PT = require('prop-types')
 var api = require('./api')
 
-var SettingsTextbox = React.createClass({
-  propTypes: {
-    name: PT.string.isRequired,
-    defaultValue: PT.string.isRequired,
-    label: PT.string.isRequired
-  },
-
-  getInitialState: function () {
-    return {
+class SettingsTextbox extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
       value: this.props.defaultValue
     }
-  },
+  }
 
-  componentDidMount: function() {
+  componentDidMount() {
     var name = this.props.name
     var defaultValue = this.props.defaultValue
-    api.settings().then( (settings) => {
+    api.settings().then((settings) => {
       var value;
       if (!settings.options) {
         value = defaultValue
       } else {
-        if(!settings.options[name]) {
+        if (!settings.options[name]) {
           value = defaultValue
         } else {
           value = settings.options[name]
         }
       }
-      this.setState({value: value})
+      this.setState({ value: value })
     })
-  },
+  }
 
-  handleChange: function(e) {
+  handleChange(e) {
     var name = this.props.name
     var value = e.target.value
-    api.setSetting(name, value).then( (result) => {
+    api.setSetting(name, value).then((result) => {
       console.log(result.updated)
       this.setState({
         value: result.settings.options[name]
       });
     });
-  },
+  }
 
-  render: function() {
+  render() {
     return (
       <p>
         <b>{this.props.label}:  </b>
         <input
           type="text"
           onChange={this.handleChange}
-          value ={this.state.value}
+          value={this.state.value}
         />
       </p>
     );
   }
-});
+}
+SettingsTextbox.propTypes = {
+  name: PT.string.isRequired,
+  defaultValue: PT.string.isRequired,
+  label: PT.string.isRequired
+}
 
 module.exports = SettingsTextbox
