@@ -1,6 +1,6 @@
 
-var React = require('react/addons')
-var cx = React.addons.classSet
+var React = require('react')
+var cx = require('classnames');
 var Link = require('react-router').Link;
 var Router = require('react-router');
 var _ = require('lodash')
