@@ -1,24 +1,30 @@
 
-var App = require('./app');
-var Post = require('./post')
-var Posts = require('./posts')
-var Page = require('./page')
-var Pages = require('./pages')
-var About = require('./about')
-var Deploy = require('./deploy')
-var Settings = require('./settings')
-var AuthSetup = require('./auth-setup')
-var Route = require('react-router').Route
+const { HashRouter, Route, Switch } = require('react-router-dom');
+const App = require('./app');
+const Post = require('./post')
+const Posts = require('./posts')
+const Page = require('./page')
+const Pages = require('./pages')
+const About = require('./about')
+const Deploy = require('./deploy')
+const Settings = require('./settings')
+const AuthSetup = require('./auth-setup')
 
 module.exports = () => {
-  return <Route handler={App}>
-    <Route name="posts" handler={Posts} path="/"/>
-    <Route name="post" handler={Post} path="/posts/:postId"/>
-    <Route name="page" handler={Page} path="/pages/:pageId"/>
-    <Route name="pages" handler={Pages} path="/pages"/>
-    <Route name="about" handler={About}/>
-    <Route name="deploy" handler={Deploy}/>
-    <Route name="settings" handler={Settings}/>
-    <Route name="auth-setup" handler={AuthSetup}/>
-  </Route>
+  return (
+    <HashRouter>
+      <App>
+        <Switch>
+          <Route name="posts" component={Posts} path="/" />
+          <Route name="post" component={Post} path="/posts/:postId" />
+          <Route name="page" component={Page} path="/pages/:pageId" />
+          <Route name="pages" component={Pages} path="/pages" />
+          <Route name="about" component={About} />
+          <Route name="deploy" component={Deploy} />
+          <Route name="settings" component={Settings} />
+          <Route name="auth-setup" component={AuthSetup} />
+        </Switch>
+      </App>
+    </HashRouter>
+  );
 }
